Made everything a float again, as the data is in floats now

+++ src/main/java/controllers/WaterData.java
@@ -2,11 +2,11 @@ package controllers;
 
 public class WaterData {
     private String datum;
-    private int TDSdata;
-    private int troebelheid;
+    private float TDSdata;
+    private float troebelheid;
     private boolean kwaliteit;
 
-    public WaterData(String datum, int TDSdata, int troebelheid, boolean kwaliteit) {
+    public WaterData(String datum, float TDSdata, float troebelheid, boolean kwaliteit) {
         this.datum = datum;
         this.TDSdata = TDSdata;
         this.troebelheid = troebelheid;
@@ -21,19 +21,19 @@ public class WaterData {
         this.datum = datum;
     }
 
-    public int getTDSdata() {
+    public float getTDSdata() {
         return TDSdata;
     }
 
-    public void setTDSdata(int TDSdata) {
+    public void setTDSdata(float TDSdata) {
         this.TDSdata = TDSdata;
     }
 
-    public int getTroebelheid() {
+    public float getTroebelheid() {
         return troebelheid;
     }
 
-    public void setTroebelheid(int troebelheid) {
+    public void setTroebelheid(float troebelheid) {
         this.troebelheid = troebelheid;
     }
